requirements.txt
Django==3.0
djangorestframework==3.14.0
pymysql==1.0.2
django-environ==0.9.0
django-cors-headers==3.13.0
pyjwt==2.5.0
pycryptodome==3.15.0
basicauth==0.4.1
twilio==7.14.2
boto3==1.24.89
boto==2.49.0
pyyaml==6.0
uritemplate==4.1.1